requirements.txt
asgiref==2.3.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.12.1
Automat==0.7.0
certifi==2018.11.29
channels==2.1.6
chardet==3.0.4
constantly==15.1.0
daphne==2.2.4
Django==2.1.4
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
oauthlib==2.1.0
PyHamcrest==1.9.0
python-dateutil==2.7.5
pytz==2018.7
pywin32==224
requests==2.21.0
requests-oauthlib==1.0.0
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24.1
zope.interface==4.6.0
